Finished all consistent hashing implementaiton methods

+++ src/main/java/com/github/arorasagar/Hasher.java
@@ -2,5 +2,5 @@ package com.github.arorasagar;
 
 public interface Hasher {
 
-    long getHash(String key, int seed);
+    long getHash(String key, long seed);
 }
